// src/types/bestkeystone/realm.ts
// Generated by https://quicktype.io

export interface Realm {
    id:                  string;
    realm_connection_id: number;
    region:              Region;
    type:                Type;
    population?:         Population;
    queue:               number;
    status:              number;
    name:                string;
    slug:                string;
    battlegroup?:        Battlegroup;
    locale:              Locale;
    timezone:            Timezone;
}

export enum Battlegroup {
    CrueltyCrueldad = "Cruelty / Crueldad",
    EmbuscadeHinterhalt = "Embuscade / Hinterhalt",
    GlutsturmEmberstorm = "Glutsturm / Emberstorm",
    Misery = "Misery",
    ReckoningAbrechnung = "Reckoning / Abrechnung",
    SturmangriffCharge = "Sturmangriff / Charge",
    VengeanceRache = "Vengeance / Rache",
    Vindication = "Vindication",
}

export enum Locale {
    DeDE = "deDE",
    EnGB = "enGB",
    EsES = "esES",
    FrFR = "frFR",
    ItIT = "itIT",
    LocaleDeDE = "de_DE",
    LocaleFrFR = "fr_FR",
    LocaleItIT = "it_IT",
    LocalePtPT = "ptPT",
    PtPT = "pt_PT",
    RuRU = "ruRU",
}

export enum Population {
    Full = "full",
    High = "high",
    Low = "low",
    Medium = "medium",
    NA = "n/a",
}

export enum Region {
    Eu = "eu",
}

export enum Timezone {
    EuropeParis = "Europe/Paris",
}

export enum Type {
    Normal = "Normal",
    Roleplaying = "Roleplaying",
    TypeNormal = "normal",
    TypeRoleplaying = "roleplaying",
}